# API/integrity-shield/src/main/kotlin/com/integrity_shield/adapter/inbound/web/dto/FileVersionDetailDto.kt
package com.integrity_shield.adapter.inbound.web.dto

import java.time.Instant

data class FileVersionDetailDto(
    val id: Long,
    val fileIdentifier: String,
    val storageVersionId: String,
    val contentHash: String,
    val leafIndex: Long,
    val uploadTimestamp: Instant? = null
)