# API/integrity-shield/src/main/kotlin/com/integrity_shield/adapter/inbound/web/dto/VerificationResponseDto.kt
package com.integrity_shield.adapter.inbound.web.dto

data class VerificationResponseDto(
    val verified: Boolean,
    val fileIdentifier: String,
    val storageVersionId: String,
    val contentHash: String
)